Issue #1111: Moved code exposing container properties from StartMojo to StartContainerExecution in order to provide container properties to the wait configuration execution.
